got one module (types) working with desired nesting
